+++ inst/app/ui.R
@@ -11,6 +11,7 @@ ui <- bslib::page_navbar(
   window_title = "courieR",
   theme = bslib::bs_theme(version = 5, preset = "shiny"),
   header = tagList(
+    shinyjs::useShinyjs(),
     tags$head(tags$link(rel = "stylesheet", href = "styles.css")),
     mod_error_reporter_ui("reporter")
   ),